refactor: 메인 페이지 스타일 조정 및 contact 페이지 삭제
