+++ src/styles/globals.css
@@ -2,25 +2,38 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-  background-color: #000;
+  background-color: rgb(35, 35, 38);
   color: white;
 }
 
 html,
 body {
   width: 100%;
-  background-color: #000;
+  background-color: rgb(35, 35, 38);
+  overflow-x: hidden;
+}
+
+.container {
+  max-width: 1200px;
+  margin: auto;
 }
 
 .nav {
   width: 100%;
   top: 0;
-  display: flex;
-  justify-content: space-between;
   padding: 2em 1em 2em 2em;
   border-bottom: 1px solid white;
   position: fixed;
   z-index: 100000;
+  background-color: rgb(35, 35, 38);
+}
+
+.nav_wrap {
+  display: flex;
+  justify-content: space-between;
+  margin: 0 auto;
+  max-width: 1200px;
+  background-color: rgb(35, 35, 38);
 }
 
 .nav a {
@@ -28,6 +41,11 @@ body {
   text-transform: uppercase;
   padding: 0 1em;
   font-weight: 700;
+  background-color: rgb(35, 35, 38);
+}
+
+.nav a:hover {
+  color: pink;
 }
 
 .container {
@@ -37,137 +55,40 @@ body {
 
 .header {
   width: 100%;
-  height: 30vh;
   padding: 2em;
-  display: flex;
   margin-top: 100px;
   align-items: center;
 }
 
-.header .col {
-  width: 60%;
+.happy {
+  color: pink;
 }
 
-.header .col:nth-child(2) {
-  width: 40%;
+.slogan {
+  font-size: 3rem;
+  position: relative;
+  text-align: right;
+  margin-bottom: 100px;
 }
 
 .header h1 {
   position: relative;
-  font-size: 80px;
+  /* font-weight: 900; */
+  font-size: 8rem;
   text-transform: uppercase;
   line-height: 1;
   margin: 0;
 }
 
-.header h2 {
-  display: none;
-}
-
-.play_wrapper {
-  position: relative;
-  width: 100%;
-  height: 80px;
-  background: url("https://images.unsplash.com/photo-1515375380578-a0587184cedd?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2899&q=80")
-    no-repeat 50% 50%;
-  background-size: cover;
-  border-radius: 50px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  padding: 0.4em;
-}
-
-.play_btn {
-  width: 70px;
-  height: 70px;
-  background: #000;
-  border-radius: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.pattern {
-  text-align: center;
-  letter-spacing: 10px;
-  color: gray;
-}
-
-.copy p {
-  font-size: 10px;
-  font-weight: 300;
-  line-height: 1.5;
-}
-
-.main {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.main img {
-  height: 59.7vh;
-}
-
-.main_wrapper {
-  position: absolute;
-  width: 0;
-  height: 100%;
-}
-
-.marquee {
-  position: absolute;
-  bottom: 0%;
-  width: 100%;
-  margin: 0 auto;
-  padding: 1em 0;
-  white-space: nowrap;
-  overflow: hidden;
-  z-index: 3;
-  background: #000;
-  display: flex;
-  align-items: center;
-}
-
-.marquee span {
-  text-transform: uppercase;
-  font-weight: 400;
-  color: #c0b1b1;
-  mix-blend-mode: normal !important;
-  font-size: 4em;
-  padding-left: -10%;
-  animation: marquee-animation 80s linear infinite;
-}
-
-@keyframes marquee-animation {
-  0% {
-    transform: translate(0, 0);
-  }
-  100% {
-    transform: translate(-100%, 0);
-  }
-}
-
-h1 {
-  font-weight: 700;
-  font-size: 5vw;
-}
-
-p {
-  margin: 1em 0 0 0;
-  font-size: 1.8vw;
-}
-
 @media (max-width: 1000px) {
+  .nickname {
+    font-size: 5vw;
+    text-align: center;
+    margin-bottom: 30px;
+  }
   .header h1 {
-    display: none;
-  }
-
-  .header h2 {
-    display: block;
-  }
-  p {
-    font-size: 3vw;
+    font-size: 14vw;
+    /* text-align: center; */
   }
 }
